requirements.txt
aioredis==1.3.1
asgiref==3.2.7
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
conda==4.8.3
conda-package-handling==1.6.0
constantly==15.1.0
cryptography==2.8
daphne==2.5.0
Django==3.0.5
django-cors-headers==3.3.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
mysqlclient==1.4.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
PySocks==1.7.1
pytz==2020.1
requests==2.22.0
ruamel-yaml==0.15.87
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
SQLAlchemy==1.3.16
sqlparse==0.3.1
tqdm==4.42.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
zope.interface==5.1.0
